refactor(smt): migrate PersistentBackend::load to RocksKVDB::new
